requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
openpyxl==3.11.0
scikit-multilearn==0.2.0
scikit-learn==1.3.2
python-jose==3.3.0
python-multipart==0.0.6
joblib==1.3.2
python-dotenv==1.0.0
flask==3.0.0
pyjwt==2.8.1
werkzeug==3.0.1
